requirements.txt
Flask
python-dotenv==0.21.0
stripe==5.5.0
decouple==0.0.7
Jinja2==3.1.2
uvicorn==0.25.0

